fix(kparser): decode the docs attribute on a struct
This commit is fixing a bug that I found while working
on a project.

In particular, the bug is related to the doc attribute
used by the compiler.

```rust
/// this is a doc comment
struct Foo {}
```

will be translated from the compiler to this

```
struct Foo {}
```

So now we should be fine, because we are adding a
attribute parsing to the struct.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>